requirements.txt
absl-py                   2.2.2
accelerate                1.7.0
aiohappyeyeballs          2.6.1
aiohttp                   3.11.18
aiosignal                 1.3.2
annotated-types           0.7.0
anyio                     4.9.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.5
async-timeout             5.0.1
attrs                     25.3.0
babel                     2.17.0
bash_kernel               0.10.0
beautifulsoup4            4.13.4
bleach                    6.2.0
blinker                   1.4
certifi                   2025.4.26
cffi                      1.17.1
charset-normalizer        3.4.1
click                     8.2.0
comm                      0.2.2
cryptography              3.4.8
datasets                  3.6.0
dbus-python               1.2.18
debugpy                   1.8.14
decorator                 5.2.1
defusedxml                0.7.1
dill                      0.3.8
distro                    1.7.0
docker-pycreds            0.4.0
einops                    0.8.1
exceptiongroup            1.2.2
executing                 2.2.0
fastjsonschema            2.21.1
filelock                  3.18.0
filetype                  1.2.0
flash-attn                2.7.2.post1
fqdn                      1.5.1
frozenlist                1.6.0
fsspec                    2025.3.0
gitdb                     4.0.12
GitPython                 3.1.44
grpcio                    1.71.0
h11                       0.16.0
httpcore                  1.0.9
httplib2                  0.20.2
httpx                     0.28.1
huggingface-hub           0.31.2
idna                      3.10
importlib-metadata        4.6.4
iniconfig                 2.1.0
iotop                     0.6
ipykernel                 6.29.5
ipython                   8.36.0
ipywidgets                8.1.6
isoduration               20.11.0
jedi                      0.19.2
jeepney                   0.7.1
Jinja2                    3.1.6
json5                     0.12.0
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2025.4.1
jupyter                   1.1.1
jupyter-archive           3.4.0
jupyter_client            8.6.3
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.12.0
jupyter-http-over-ws      0.0.8
jupyter-lsp               2.2.5
jupyter_server            2.15.0
jupyter_server_terminals  0.5.3
jupyterlab                4.4.1
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.14
keyring                   23.5.0
launchpadlib              1.10.16
lazr.restfulclient        0.14.4
lazr.uri                  1.0.6
Markdown                  3.8
MarkupSafe                3.0.2
matplotlib-inline         0.1.7
mistune                   3.1.3
more-itertools            8.10.0
mpmath                    1.3.0
multidict                 6.4.3
multiprocess              0.70.16
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nbzip                     0.1.0
nest-asyncio              1.6.0
networkx                  3.4.2
notebook                  7.4.1
notebook_shim             0.2.4
numpy                     1.25.2
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         8.9.2.26
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-nccl-cu12          2.18.1
nvidia-nvjitlink-cu12     12.9.41
nvidia-nvtx-cu12          12.1.105
oauthlib                  3.2.0
overrides                 7.7.0
packaging                 25.0
pandas                    2.2.3
pandocfilters             1.5.1
parso                     0.8.4
pexpect                   4.9.0
pip                       22.0.2
platformdirs              4.3.7
pluggy                    1.5.0
prometheus_client         0.21.1
prompt_toolkit            3.0.51
propcache                 0.3.1
protobuf                  6.30.2
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pyarrow                   20.0.0
pycparser                 2.22
pydantic                  2.11.4
pydantic_core             2.33.2
Pygments                  2.19.1
PyGObject                 3.42.1
PyJWT                     2.3.0
pyparsing                 2.4.7
pytest                    8.3.5
python-apt                2.4.0+ubuntu4
python-dateutil           2.9.0.post0
python-json-logger        3.3.0
pytz                      2025.2
PyYAML                    6.0.2
pyzmq                     26.4.0
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.24.0
safetensors               0.5.3
SecretStorage             3.3.1
Send2Trash                1.8.3
sentry-sdk                2.28.0
setproctitle              1.3.6
setuptools                59.6.0
six                       1.16.0
smmap                     5.0.2
sniffio                   1.3.1
soupsieve                 2.7
stack-data                0.6.3
supervisor                4.2.1
sympy                     1.14.0
tensorboard               2.19.0
tensorboard-data-server   0.7.2
terminado                 0.18.1
tinycss2                  1.4.0
tokenizers                0.21.1
tomli                     2.2.1
torch                     2.1.0
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
transformers              4.51.3
triton                    2.1.0
types-python-dateutil     2.9.0.20241206
typing_extensions         4.13.2
typing-inspection         0.4.0
tzdata                    2025.2
uri-template              1.3.0
urllib3                   2.4.0
uv                        0.7.4
wadllib                   1.3.6
wandb                     0.19.11
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.1.3
wheel                     0.37.1
widgetsnbextension        4.0.14
xxhash                    3.5.0
yarl                      1.20.0
zipp                      1.0.0
